video_tools/requirements.txt
fastapi
uvicorn
requests
moviepy 
openai-whisper
edge-tts
torch
python-dotenv
pydantic-settings
python-multipart
